Vendor x86_64 crate with u64 replaced with usize

+++ src/mem/addr_space.rs
@@ -31,7 +31,7 @@ lazy_static! {
 
         AddrSpace {
             table: RwSpinLock::new(unsafe {
-                OffsetPageTable::new(&mut *table_virt.as_mut_ptr(), super::PHYS_OFFSET as u64)
+                OffsetPageTable::new(&mut *table_virt.as_mut_ptr(), super::PHYS_OFFSET)
             }),
         }
     };
